# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CamMonSys API
  version: 1.0.0
  description: Role-based camera monitoring REST and Server-Sent Events API.
servers:
  - url: /
    description: Same-origin application server
security:
  - cookieAuth: []
...
paths:
...
  /api/devices/{deviceId}/events:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    post:
      tags: [Events]
      summary: Create a simulated camera event
      description: Admin only. Requires ownership of the active publication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [publishingSessionId, type]
              properties:
                publishingSessionId: { type: string, format: uuid }
                type: { $ref: '#/components/schemas/EventType' }
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [publishingSessionId, type]
              properties:
                publishingSessionId: { type: string, format: uuid }
                type: { $ref: '#/components/schemas/EventType' }
                snapshot:
                  type: string
                  format: binary
                  description: Optional JPEG snapshot, at most 2 MiB.
      responses:
        '201':
          description: Event persisted.
          content:
            application/json:
              schema:
                type: object
                required: [event]
                properties:
                  event: { $ref: '#/components/schemas/CreatedEvent' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: Snapshot exceeds the 2 MiB limit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '409':
          description: Publication is inactive.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  /api/events:
    get:
      tags: [Events]
      summary: List monitoring events
      description: Admin/Monitor can optionally filter by device. Responders must use assignedTo=me and receive unresolved assignments only.
      parameters:
        - name: deviceId
          in: query
          schema: { type: string, format: uuid }
        - name: assignedTo
          in: query
          schema: { type: string, const: me }
      responses:
        '200':
          description: Authorized events, up to 100.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/MonitoringEvent' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/events/{eventId}:
    parameters:
      - { $ref: '#/components/parameters/EventId' }
    get:
      tags: [Events]
      summary: Get monitoring event detail
      description: Admin and Monitor can view any event. A Responder can view only their own unresolved assignment.
      responses:
        '200':
          description: Complete event lifecycle detail.
          content:
            application/json:
              schema:
                type: object
                required: [event]
                properties:
                  event: { $ref: '#/components/schemas/MonitoringEvent' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/events/{eventId}/snapshot:
    parameters:
      - { $ref: '#/components/parameters/EventId' }
    get:
      tags: [Events]
      summary: Get an event snapshot
      description: Returns the JPEG snapshot to Admin/Monitor, or to the assigned Responder while the event is unresolved.
      responses:
        '200':
          description: Snapshot image.
          headers:
            Cache-Control:
              schema: { type: string, const: 'private, no-store' }
            X-Content-Type-Options:
              schema: { type: string, const: nosniff }
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Event, accessible assignment, or snapshot not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/events/{eventId}/status:
    parameters:
      - { $ref: '#/components/parameters/EventId' }
    patch:
      tags: [Events]
      summary: Advance an event status
      description: Admin, Monitor, or assigned Responder. Allows only OPEN to ACKNOWLEDGED to RESOLVED.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  additionalProperties: false
                  required: [status]
                  properties:
                    status: { type: string, const: ACKNOWLEDGED }
                - type: object
                  additionalProperties: false
                  required: [status]
                  properties:
                    status: { type: string, const: RESOLVED }
                    completionNote: { type: string, minLength: 1, maxLength: 1000 }
      responses:
        '200':
          description: Status updated.
          content:
            application/json:
              schema:
                type: object
                required: [event]
                properties:
                  event: { $ref: '#/components/schemas/UpdatedEvent' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Invalid current status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/events/{eventId}/assignment:
    parameters:
      - { $ref: '#/components/parameters/EventId' }
    patch:
      tags: [Events]
      summary: Assign, reassign, or unassign a Responder
      description: Admin or Monitor only. Resolved events cannot be changed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  additionalProperties: false
                  required: [responderId, instructions]
                  properties:
                    responderId: { type: string, format: uuid }
                    instructions: { type: string, minLength: 1, maxLength: 1000 }
                - type: object
                  additionalProperties: false
                  required: [responderId]
                  properties:
                    responderId: { type: 'null' }
                    instructions: { type: 'null' }
      responses:
        '200':
          description: Assignment updated.
          content:
            application/json:
              schema:
                type: object
                required: [event]
                properties:
                  event: { $ref: '#/components/schemas/EventAssignment' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Event or Responder does not exist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '409':
          description: Event is resolved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cammon_token
      description: HttpOnly signed JWT backed by a revocable PostgreSQL session.
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    EventId:
      name: eventId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    InvalidInput:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Authentication is missing, invalid, expired, or revoked.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Forbidden:
      description: The user lacks permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    EventType:
      type: string
      enum: [TEST_ALERT, MOTION, BED_EXIT]
    EventStatus:
      type: string
      enum: [OPEN, ACKNOWLEDGED, RESOLVED]
...
    CreatedEvent:
      type: object
      additionalProperties: false
      required: [id, device_id, type, status, created_at, assigned_to_id, snapshot_available, snapshot_captured_at]
      properties:
        id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/EventType' }
        status: { $ref: '#/components/schemas/EventStatus' }
        created_at: { type: string, format: date-time }
        assigned_to_id: { $ref: '#/components/schemas/NullableUuid' }
        snapshot_available: { type: boolean }
        snapshot_captured_at: { $ref: '#/components/schemas/NullableDateTime' }
    MonitoringEvent:
      type: object
      additionalProperties: false
      required: [id, device_id, device_name, device_location, type, status, created_at, assigned_to_id, assigned_to_name, assigned_by_name, instructions, completion_note, snapshot_available, snapshot_captured_at, acknowledged_at, acknowledged_by_name, resolved_at, resolved_by_name]
      properties:
        id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        device_name: { type: string }
        device_location: { type: string }
        type: { $ref: '#/components/schemas/EventType' }
        status: { $ref: '#/components/schemas/EventStatus' }
        created_at: { type: string, format: date-time }
        assigned_to_id: { $ref: '#/components/schemas/NullableUuid' }
        assigned_to_name: { $ref: '#/components/schemas/NullableString' }
        assigned_by_name: { $ref: '#/components/schemas/NullableString' }
        instructions: { $ref: '#/components/schemas/NullableString' }
        completion_note: { $ref: '#/components/schemas/NullableString' }
        snapshot_available: { type: boolean }
        snapshot_captured_at: { $ref: '#/components/schemas/NullableDateTime' }
        acknowledged_at: { $ref: '#/components/schemas/NullableDateTime' }
        acknowledged_by_name: { $ref: '#/components/schemas/NullableString' }
        resolved_at: { $ref: '#/components/schemas/NullableDateTime' }
        resolved_by_name: { $ref: '#/components/schemas/NullableString' }
    UpdatedEvent:
      type: object
      additionalProperties: true
      required: [id, device_id, type, status, created_at, assigned_to_id]
      properties:
        id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/EventType' }
        status: { $ref: '#/components/schemas/EventStatus' }
        created_at: { type: string, format: date-time }
        assigned_to_id: { $ref: '#/components/schemas/NullableUuid' }
        completion_note: { $ref: '#/components/schemas/NullableString' }
    EventAssignment:
      type: object
      additionalProperties: false
      required: [id, device_id, status, assigned_to_id, assigned_by_id, instructions]
      properties:
        id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/EventStatus' }
        assigned_to_id: { $ref: '#/components/schemas/NullableUuid' }
        assigned_by_id: { $ref: '#/components/schemas/NullableUuid' }
        instructions: { $ref: '#/components/schemas/NullableString' }
...
    NullableUuid:
      oneOf:
        - { type: string, format: uuid }
        - { type: 'null' }
    NullableString:
      oneOf:
        - { type: string }
        - { type: 'null' }
    NullableDateTime:
      oneOf:
        - { type: string, format: date-time }
        - { type: 'null' }
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...